refactor: 확장성을 위해서 모달 구조를 변경한다
