+++ src/js/modals/StationModal.js
@@ -1,37 +1,39 @@
-import { ALERT_MESSAGE, ID_SELECTOR, REQUEST_URL } from '../constants';
+import Component from '../components/Component';
+import { ALERT_MESSAGE, ID_SELECTOR, KEYWORD, REQUEST_URL } from '../constants';
 import STATION_TEMPLATE from '../templates/stationTemplate';
-import { openModal, closeModal } from '../utils/DOM';
+import { closeModal } from '../utils/DOM';
 import { fetchStationNameRevision } from '../utils/fetch';
 import { loadStationList } from '../utils/loadByAJAX';
 import $ from '../utils/querySelector';
-import Component from './Component';
+import Modal from './Modal';
 
-class StationModalComponent extends Component {
-  modalState;
+class StationModal extends Modal {
+  constructor(props) {
+    super(props);
 
+    this._router = {
+      [KEYWORD.REVISION]: new StationRevisionComponent({
+        accessTokenState: this.props.accessTokenState,
+        stationsState: this.props.stationsState,
+      }),
+    };
+  }
+}
+
+class StationRevisionComponent extends Component {
   constructor(props) {
     super(props);
   }
 
-  initLoad() {
-    $(`#${ID_SELECTOR.MODAL}`).innerHTML = STATION_TEMPLATE.MODAL;
-  }
-
   initEvent() {
-    //TODO : dimmed 부분을 누르면 모달창 닫히게
     $(`#${ID_SELECTOR.STATION_MODAL_FORM}`).addEventListener(
       'submit',
       this.#onRevisionSubmit
     );
   }
 
-  renderStationModal(target) {
-    const stationId = target.dataset.id;
-    const stationName = target.dataset.name;
-
-    $(`#${ID_SELECTOR.STATION_MODAL_FORM_INPUT}`).value = stationName;
-    $(`#${ID_SELECTOR.STATION_MODAL_FORM_INPUT}`).dataset.id = stationId;
-    openModal();
+  render() {
+    $(`#${ID_SELECTOR.MODAL}`).innerHTML = STATION_TEMPLATE.MODAL;
   }
 
   #onRevisionSubmit = async event => {
@@ -47,6 +49,7 @@ class StationModalComponent extends Component {
     const accessToken = this.props.accessTokenState.Data;
 
     try {
+      // TODO: fetchStationNameRevision 이름 변경하기
       await fetchStationNameRevision(url, { bodyData, accessToken });
       alert(ALERT_MESSAGE.STATION_NAME_REVISION_SUCCESS);
       closeModal();
@@ -61,4 +64,4 @@ class StationModalComponent extends Component {
   };
 }
 
-export default StationModalComponent;
+export default StationModal;
